pyproject.toml
[project]
name = "viewshed"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100.0",
    "httpx>=0.28.1",
    "matplotlib>=3.5.0",
    "numba>=0.65.1",
    "numpy>=1.21.0",
    "pyyaml>=6.0",
    "shapely>=2.0.0",
    "uvicorn[standard]>=0.23.0",
]
